+++ src/Appwrite/Platform/Modules/Databases/Http/TablesDB/Transactions/Delete.php
@@ -30,7 +30,7 @@ class Delete extends TransactionsDelete
             ->setHttpPath('/v1/tablesdb/transactions/:transactionId')
             ->desc('Delete transaction')
             ->groups(['api', 'database', 'transactions'])
-            ->label('scope', 'rows.write')
+            ->label('scope', ['documents.write', 'rows.write'])
             ->label('resourceType', RESOURCE_TYPE_DATABASES)
             ->label('sdk', new Method(
                 namespace: 'tablesDB',